artifacts/api-server/requirements.txt
django==5.2.12
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
gunicorn==23.0.0
